fix(mockworld): break src->tests import + clean arch-doc generator (PR A pass-1 follow-ups)
ADR-0051 fresh-eyes review pass #1 surfaced 1 Critical + 2 Important
that per-task reviews missed at the cumulative-diff level:

C1 (Critical). src/mockworld/fakes/fake_github.py and fake_llm.py
imported from tests.conftest, breaking sandbox_main in PR B's docker
container which doesn't COPY tests/. Relocated PRInfoFactory,
PlanResultFactory, ReviewResultFactory, TriageResultFactory,
WorkerResultFactory to src/mockworld/fakes/_factories.py; re-exported
from tests/conftest.py for back-compat. Fakes now resolve under
PYTHONPATH=src alone.

I1 (Important). docs/arch/generated/mockworld.md invented IssuePort,
IssueRecordPort, IssueSummaryPort etc. by naive Fake<Stem>->StemPort
matching against nested dataclasses (FakeIssue, FakePR, FakeIssueRecord,
FakeIssueSummary). Filtered the discovery logic in
src/arch/extractors/mockworld.py to gate on _is_fake_adapter=True via
AST inspection, and propagated the marker to the 9 top-level Fakes that
lacked it (FakeBeads, FakeClock, FakeDocker, FakeFS, FakeGit, FakeHTTP,
FakeSentry, FakeSubprocessRunner, FakeWikiCompiler). Regenerated docs.

I2 (Important). build_services()'s cast(ConcreteAdapter, override)
pattern is duck-safe at runtime but type-checker blind. Expanded
the comment at each cast site (workspaces, prs, fetcher, store) to
explicitly call out the pyright blind spot and reference Task 1.9's
full widening as the mitigation path.

Bonus M1 (parity test): tests/scenarios/test_sandbox_parity.py was
silently pytest.skip()'ing on KeyError, hiding scenarios where loop
catalog drifts. Switched to pytest.xfail so missing-loop drift is
visible in test summary.
