+++ src/arch/generators/mockworld_map.py
@@ -7,9 +7,11 @@ from arch._models import MockWorldMap
 _HEADER = "# MockWorld Map\n\n"
 _PREAMBLE = (
     "<!-- generated by arch.generators.mockworld_map; do not hand-edit -->\n\n"
-    "All fakes under `src/mockworld/fakes/`, the `*Port` each implements "
-    "(by name match), and the scenarios that wire them. Per ADR-0022 "
-    "(MockWorld) and ADR-0047 (fake-adapter contract testing).\n\n"
+    "All Fake adapters under `src/mockworld/fakes/` (classes with "
+    "``_is_fake_adapter = True``; nested-record dataclasses are filtered "
+    "out), the `*Port` each implements (by name match), and the scenarios "
+    "that wire them. Per ADR-0022 (MockWorld) and ADR-0047 "
+    "(fake-adapter contract testing).\n\n"
 )
 _FOOTER = "\n\n{{ARCH_FOOTER}}\n"
